IOW-135 - statistical gw daily data
