+++ src/main/java/gov/usgs/wma/waterdata/geojson/Geometry.java
@@ -1,4 +1,4 @@
-package gov.usgs.wma.waterdata.location.geojson;
+package gov.usgs.wma.waterdata.geojson;
 
 import java.math.BigDecimal;
 
@@ -9,21 +9,11 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @JsonPropertyOrder(value={"type","coordinates"})
 @Schema(description="A Geometry object represents points, curves, and surfaces in coordinate space.")
 public class Geometry {
-	@Schema(hidden=true)
-	private BigDecimal longitude;
-	@Schema(hidden=true)
-	private BigDecimal latitude;
 	@Schema(description="The type of this Geometry object - Currently only 'Point' (longitude, latitude) objects are produced.", example="Point")
 	public String getType() {
 		return "Point";
 	}
 	public BigDecimal[] getCoordinates() {
-		return new BigDecimal[] {longitude,latitude};
-	}
-	public void setLongitude(BigDecimal longitude) {
-		this.longitude = longitude;
-	}
-	public void setLatitude(BigDecimal latitude) {
-		this.latitude = latitude;
+		return new BigDecimal[] {new BigDecimal(-102.480479), new BigDecimal(35.667543)};
 	}
 }
